Add starter models, routers, controllers, test data, tests, and app.js
This commit adds all the starter code needed to begin the project.
This includes all the models for the Mongoose ORM to the database.
In addition, routers that define all the endpoints have been created
with TODOs above all functions that have not been implemented.
Controllers have also been created with TODOs above all functions
that need to be implemented, along with a brief comment about
what end points in the router some controller functions will
correspond to.

The skeleton for integration tests have been created so that they can be
implemented as the project moves further along.

Test data has also been provided as a template example in case
additional data needs to be added to the database for testing purposes.
Keep in mind that using MongoDB Atlas tool requires a slightly different
format for inserting documents, specifically related to how parentheses
are used.