Update README with dev environment setup info
I've added steps for setting up a local copy of this project for anyone interested in submitting pull requests in the future. These steps can be reorganized, reformatted, and/or reworded in the future as we see fit. The required tools before setting up the project (npm, node.js, etc.) have also been listed.